Added package com.clouway.exceptions.linkedlistexceptions
Class List represents a linked list with methods for adding element at the end of the list, removing element from the end of the list and printing all elements of the list. Added two new classes - FullListException and EmptyListException which are used when we try to add/remove element from the list.